feat: category 생성/ CSS 작업
1) 카테고리 생성 작업 구현
2) 관련 페이지 css 작업

+++ src/main/java/com/skhuedin/skhuedin/domain/Category.java
@@ -23,10 +23,10 @@ public class Category extends BaseEntity {
 
     private String name;
 
-    private Integer weight;
+    private Long weight;
 
     @Builder
-    public Category(String name, Integer weight) {
+    public Category(String name, Long weight) {
         this.name = name;
         this.weight = weight;
     }
